Teams-app/requirements.txt
botbuilder-core
fastapi[all]
aiohttp
agno
docker
groq
lancedb
fastembed
sqlalchemy